Move SignaturesReportController out of the admin panel
* require authentication from within the controller, instead of relying on
  the auth wrapper in the :admin routes
