+++ config/routes.rb
@@ -9,7 +9,6 @@ Rails.application.routes.draw do
       resources :shifts
       resources :report_recipients
       # resources :shift_events ### not needed for this app
-      resources :signatures_report, only: [:index], defaults: { format: :csv }
 
       root to: 'shifts#index'
     end
@@ -18,4 +17,5 @@ Rails.application.routes.draw do
   root to: 'check_ins#new'
 
   resources :check_ins, only: [:new, :create]
+  resources :signatures_report, only: [:index], defaults: { format: :csv }
 end
